PA-BD/biblioteca/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9
drf-yasg>=1.21.6